setting the right value for signature
android calls this value "signature" but it really is a public key to a
private key that signed the app. this means this value does not need to
be updated as long as the signing key doesn't change and here in this
case it's also equal between the two apps.